+++ tests/mapping/test_MappingTransformer.py
@@ -11,7 +11,6 @@ from pandas.api.types import (
 
 import tests.test_data as d
 import tubular
-from tubular.base import ReturnKeyDict
 from tubular.mapping import MappingTransformer
 
 
@@ -33,7 +32,7 @@ class TestInit:
         call_kwargs = call_args[1]
 
         expected_kwargs = {
-            "mappings": {"a": ReturnKeyDict({"a": 1})},
+            "mappings": {"a": {"a": 1}},
             "verbose": True,
             "copy": True,
         }
@@ -47,28 +46,6 @@ class TestInit:
         assert (
             expected_pos_args == call_pos_args
         ), "unexpected positional args in BaseMappingTransformer.__init__ call"
-
-    def test_mapping_arg_conversion(self):
-        """Test that sub dict in mappings arg are converted into ReturnKeyDict objects."""
-        mappings = {
-            "a": {"a": 1, "b": 2},
-            "b": {1: 4.5, 2: 3.1},
-            "c": {False: None, True: 1},
-        }
-
-        expected_mappings = {
-            "a": ReturnKeyDict(mappings["a"]),
-            "b": ReturnKeyDict(mappings["b"]),
-            "c": ReturnKeyDict(mappings["c"]),
-        }
-
-        x = MappingTransformer(mappings=mappings)
-
-        ta.equality.assert_equal_dispatch(
-            expected_mappings,
-            x.mappings,
-            "mappings attribute not correctly converted sub dicts to ReturnKeyDict",
-        )
 
     def test_mapping_non_dict_item_error(self):
         """Test an exception is raised if mappings contains non-dict values."""
@@ -186,8 +163,8 @@ class TestTransform:
         }
 
         preserve_original_value_mapping = {
-            "a": ReturnKeyDict(mapping["a"]),
-            "b": ReturnKeyDict(mapping["b"]),
+            "a": dict(mapping["a"]),
+            "b": dict(mapping["b"]),
         }
 
         x = MappingTransformer(mappings=mapping)