+++ CHANGELOG.rst
@@ -28,10 +28,14 @@ Changed
 ^^^^^^^
 - minor changes to comply with flake8_bugbear (B) ruff rules `#131 <https://github.com/lvgig/tubular/pull/131>`_
 - minor changes to comply with flake8_datetimez (DTZ) ruff rules `#132 <https://github.com/lvgig/tubular/pull/132>`_
+- BaseMappingTransformerMixin chnaged to use Dataframe.replace rather than looping over columns `#135 <https://github.com/lvgig/tubular/pull/135>`_
+- MeanResponseTransformer.map_imputer_values() added to decouple from BaseMappingTransformerMixin `#135 <https://github.com/lvgig/tubular/pull/135>`_
 
 Removed
 ^^^^^^^
 - removed some unnescessary implementation tests `#130 <https://github.com/lvgig/tubular/pull/130>`_
+- ReturnKeyDict class removed `#135 <https://github.com/lvgig/tubular/pull/135>`_
+